refactor(tests): group drum and clip suites into their own folders
Both directories were at the 12-item folder limit with more drum-pad
coverage still to come. Drum e2e tests move to e2e/mcp/device/drum/ and
clip duplicate tests to tests/clip/, mirroring the helpers/clip/ split
they already sit beside.

Also lifts createMidiTrack into mcp-test-helpers, since every drum suite
needs it and each was about to grow its own copy.
