+++ e2e/mcp/device/drum/nested-rack-drum-detection.test.ts
@@ -28,8 +28,8 @@ import {
   setConfig,
   setupMcpTestContext,
   sleep,
-} from "../mcp-test-helpers.ts";
-import { RACKS_TEST_PATH } from "./helpers/racks-test-helpers.ts";
+} from "../../mcp-test-helpers.ts";
+import { RACKS_TEST_PATH } from "../helpers/racks-test-helpers.ts";
 
 const ctx = setupMcpTestContext({ once: true, liveSetPath: RACKS_TEST_PATH });
