+++ e2e/mcp/device/drum/ppal-create-device-drum-kit.test.ts
@@ -18,7 +18,7 @@ import {
   parseToolResultWithWarnings,
   setupMcpTestContext,
   sleep,
-} from "../mcp-test-helpers";
+} from "../../mcp-test-helpers";
 
 const ctx = setupMcpTestContext({ once: true });
