+++ src/tools/actions/duplicate/tests/clip/duplicate-take-lane.test.ts
@@ -5,7 +5,7 @@
 
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import { livePath } from "#src/shared/live-api-path-builders.ts";
-import "./duplicate-mocks-test-helpers.ts";
+import "../duplicate-mocks-test-helpers.ts";
 import {
   lookupMockObject,
   registerMockObject,
